# docs/phase5/logs/openapi-v1-phase5.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "DerasaX.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/vision/sessions": {
      "post": {
        "tags": [
          "ClassroomVision"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.VisionDto.StartVisionSessionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.VisionDto.StartVisionSessionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.VisionDto.StartVisionSessionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "ClassroomVision"
        ],
        "parameters": [
          {
            "name": "SchoolClassId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/DerasaX.Domain.Enums.VisionSessionStatus"
            }
          },
          {
            "name": "From",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "To",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/vision/sessions/{id}": {
      "get": {
        "tags": [
          "ClassroomVision"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/vision/sessions/{id}/end": {
      "post": {
        "tags": [
          "ClassroomVision"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/vision/sessions/{id}/analyze": {
      "post": {
        "tags": [
          "ClassroomVision"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.VisionDto.AnalyzeFrameDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.VisionDto.AnalyzeFrameDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.VisionDto.AnalyzeFrameDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/vision/sessions/{id}/frames": {
      "get": {
        "tags": [
          "ClassroomVision"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/vision/sessions/{id}/candidates": {
      "get": {
        "tags": [
          "ClassroomVision"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ReviewStatus",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/DerasaX.Domain.Enums.CandidateReviewStatus"
            }
          },
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/vision/sessions/{id}/summary": {
      "get": {
        "tags": [
          "ClassroomVision"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/vision/candidates/{candidateId}/confirm": {
      "post": {
        "tags": [
          "ClassroomVision"
        ],
        "parameters": [
          {
            "name": "candidateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.VisionDto.ConfirmCandidateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.VisionDto.ConfirmCandidateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.VisionDto.ConfirmCandidateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/vision/candidates/{candidateId}/reject": {
      "post": {
        "tags": [
          "ClassroomVision"
        ],
        "parameters": [
          {
            "name": "candidateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.VisionDto.RejectCandidateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.VisionDto.RejectCandidateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.VisionDto.RejectCandidateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/vision/candidates/{candidateId}/override": {
      "post": {
        "tags": [
          "ClassroomVision"
        ],
        "parameters": [
          {
            "name": "candidateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.VisionDto.OverrideCandidateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.VisionDto.OverrideCandidateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.VisionDto.OverrideCandidateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/vision/enrollments": {
      "post": {
        "tags": [
          "ClassroomVision"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.VisionDto.EnrollFaceDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.VisionDto.EnrollFaceDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.VisionDto.EnrollFaceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "ClassroomVision"
        ],
        "parameters": [
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/vision/enrollments/{enrollmentId}/asset": {
      "post": {
        "tags": [
          "ClassroomVision"
        ],
        "parameters": [
          {
            "name": "enrollmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "File": {
                    "type": "string",
                    "format": "binary"
                  },
                  "ConsentObtained": {
                    "type": "boolean"
                  },
                  "ConsentReference": {
                    "type": "string"
                  },
                  "RetentionDays": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              },
              "encoding": {
                "File": {
                  "style": "form"
                },
                "ConsentObtained": {
                  "style": "form"
                },
                "ConsentReference": {
                  "style": "form"
                },
                "RetentionDays": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/vision/enrollments/{enrollmentId}/asset/download": {
      "get": {
        "tags": [
          "ClassroomVision"
        ],
        "parameters": [
          {
            "name": "enrollmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/enrollments": {
      "get": {
        "tags": [
          "Enrollments"
        ],
        "parameters": [
          {
            "name": "SchoolClassId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "StudentId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/DerasaX.Domain.Enums.EnrollmentStatus"
            }
          },
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Enrollments"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.AcademicDto.AddEnrollmentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.AcademicDto.AddEnrollmentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.AcademicDto.AddEnrollmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DerasaX.Application.Dto.AcademicDto.AddEnrollmentDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string",
            "nullable": true
          },
          "schoolClassId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DerasaX.Application.Dto.VisionDto.AnalyzeFrameDto": {
        "type": "object",
        "properties": {
          "imageBase64": {
            "type": "string",
            "nullable": true
          },
          "frameIndex": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "captureLabel": {
            "type": "string",
            "nullable": true
          },
          "wantEngagement": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DerasaX.Application.Dto.VisionDto.ConfirmCandidateDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "remember": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "DerasaX.Application.Dto.VisionDto.EnrollFaceDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string",
            "nullable": true
          },
          "externalLabelId": {
            "type": "string",
            "nullable": true
          },
          "displayLabel": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DerasaX.Application.Dto.VisionDto.OverrideCandidateDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DerasaX.Application.Dto.VisionDto.RejectCandidateDto": {
        "type": "object",
        "properties": {
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DerasaX.Application.Dto.VisionDto.StartVisionSessionDto": {
        "type": "object",
        "properties": {
          "schoolClassId": {
            "type": "string",
            "nullable": true
          },
          "subjectId": {
            "type": "string",
            "nullable": true
          },
          "lessonId": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "sessionDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "recognitionThreshold": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DerasaX.Domain.Enums.CandidateReviewStatus": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "DerasaX.Domain.Enums.EnrollmentStatus": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "DerasaX.Domain.Enums.VisionSessionStatus": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}